move StorageViewArrayIterator out of item
